feat: useMultipleChoice, useTextAnswer 훅 분리
- QnABox의 객관식 문항 관련 코드는 useMultipleChoice, 서술형 관련 코드는 useTextAnswer 훅으로 분리
- upatedAnswerMap -> upatedAnswerMap 네이밍 변경
