+++ frontend/src/hooks/review/writingCardForm/index.ts
@@ -1,3 +1,5 @@
 export { default as useReviewerAnswer } from './useReviewerAnswer';
 export { default as useSlideWidth } from './useSlideWidth';
 export { default as useCurrentCardIndex } from './useCurrentCardIndex';
+export { default as useMultipleChoice } from './useMultipleChoice';
+export { default as useTextAnswer } from './useTextAnswer';
